predabst: delete (#ifdefd out) code that avoid using push/pop; incremental computation of the Cartesian product really needs push/pop in order to work efficiently.
Don't apply a rule to nodes that were created later than the node that is currently being processed, to avoid duplicate work.  Also avoid duplicate work when a rule takes the same predicate in multiple positions in its body.

Revert to the more logical order for choosing the nodes for the body predicates of a rule (i.e. left to right, not right to left).