predabst: use iz3interpolate to solve clauses.  Unfortunately, this doesn't appear to be behaving correctly at the moment.
Fix a bug in clause generation where clauses would refer to other clauses that didn't exist (because that part of the tree wasn't in the unsat core).  This is a pretty ugly fix; I'll improve later.  Add a new test case that should catch this bug.

Get rid of root_name, and just hard-code 0.  The first clause in the clauses list  is now assumed to be the root.  Get rid of any notion that the index in the clauses list is the same as the head name.

Generalize replace_preds.

Add some code to make_tests to generate a large random problem programmatically.  This is a direction I'd like to explore further.