predabst: stop using iz3interpolate for the time being; instead, handle the case where the unsat proof is non-Farkas by translating to the dual problem and solving for the coefficients explicitly.
Note: this will fail for template@head rules where the template instance is not atomic: we need to convert to DNF.

Perform the theory rewrite on the whole predicate (not just its LHS), in order, for example, to reduce 0 = 0 to true, and therefore to avoid adding worthless/duplicate predicates.  (Actually, this doesn't work as well as I thought: theory rewriting doesn't seem to normalize all equivalent inequalities.  Investigation ongoing.)

Use maybe_add_pred for predicates from the input file as well, so that they get rewritten in the same way.  This ensures that (a) we strip duplicate predicates from the input, and (b) generated predicates are correctly identified as duplicates of input predicates whenever they are.