predabst: support defining argument names with, for example:
  (assert (forall ((x Int) (y Int)) (=> (and (__name__foo x) (__name__bar y)) (__names__p x y))))

if the names of the arguments x and y of p(x, y) are "foo" and "bar" respectively.

Fix a bug where we assumed that all arguments with a single name would be of the same type; now the type is effectively treated as being part of the name when considering whether to translate predicates from one symbol to another.